Made Session and User Dao for full functionality. Added the stored procedure for Users.

+++ kickIn_26/src/main/java/nl/utwente/di/team26/Product/model/Map/Map.java
@@ -33,7 +33,7 @@ public class Map implements Serializable {
 
     }
 
-    public Map(int mapIdIn) {
+    public Map(long mapIdIn) {
         this.mapId = mapIdIn;
     }
 
@@ -42,7 +42,7 @@ public class Map implements Serializable {
         this.description = description;
     }
 
-    public Map(int mapId, String name, String description) {
+    public Map(long mapId, String name, String description) {
         this.mapId = mapId;
         this.name = name;
         this.description = description;
